+++ CHANGELOG.md
@@ -35,6 +35,8 @@ change log follows the conventions of
 
 * pink.filter
   * zdf-ladder - a Zero-delay feedback Moog Ladder filter (4-pole 24db/oct)
+  * zdf-1pole - zero-delay feedback, 1-pole (6 dB/oct) state variable filter
+    (returns low-pass and high-pass signals)
   * zdf-2pole - zero-delay feedback, 2-pole (12 dB/oct) state variable filter
     (returns low-pass, band-pass, and high-pass signals)